pyproject.toml
[tool.poetry]
name = "chat-langchain"
version = "0.1.1"
description = ""
authors = ["SN <6432132+samnoyes@users.noreply.github.com>"]
readme = "README.md"
packages = [{ include = "backend" }]

[tool.poetry.dependencies]
python = "^3.11"
langchain = ">=0.3.0,<0.4.0"
langsmith = "0.4.9" 
langchain-google-genai = ">=2.1.2,<3.0.0"
langchain-core = ">=0.3.10,<0.4.0"
langchain-community = ">=0.3.0,<0.4.0"
langchain-openai = ">=0.3.12,<0.4.0"
langchain-anthropic = ">=0.3.10,<0.4.0"
langchain-weaviate = ">=0.0.3,<0.1.0"
langgraph = ">=0.4.5"
beautifulsoup4 = "^4.12.2"
weaviate-client = "^4.0.0"
lxml = "^4.9.3"
voyageai = "^0.1.4"
pillow = "^10.2.0"
# do not remove this, it's used for indexing
psycopg2-binary = "^2.9.9"
langchainhub = "^0.1.21"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
langgraph-sdk = ">=0.2.0,<0.3.0"
pytest = "^7.3.0"
ruff = "^0.2.2"
pandas = "^2.2.2"
langgraph-cli = {extras = ["inmem"], version = "^0.4.2"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
